Feature: New job-submit plugin added at ../plugins/job_submit/sflex
sflex job-submit plugin is like a toy example to understand how to add new plugin to slurm codebase. Job-submit plugin will be called for each job-step right after
request is recived by slurmctld. sflex plugin will check if it is a gres type job and modifies it to --gres=mps:10. The precise steps followed to add this plugin are:
	-create folder src/plugins/job_submit/sflex , inside create job_submit_sflex.c(the name of the file should start with job_submit job_submit*.c) , Makefile.am
	-Job_submit_sflex.c should have your plugin functions including init,fini,job_submit etc.,
	-Parent directory makefile should be informed about this new folder,so ew entry for sflex inside job_submit/Makefile.am should be created
	-Run autoconf in to slurm ../slurm-23.02.2 directory, this will create necessary Makefile,Makefile.in in the /sflex, generates configure binary
	-Run ./configure --prefix=/usr --sysconfdir=/etc/slurm --enable-pam --with-pam_dir=/lib/x86_64-linux-gnu/security/ --without-shared-libslurm
	-make \ sudo make install will add new features.(Refer to google doc for detatiled summary)